expanded tests/controllers

+++ programs/clearing_house/src/state/insurance_fund_stake.rs
@@ -8,6 +8,6 @@ pub struct InsuranceFundStake {
     pub bank_index: u64,
     pub lp_shares: u128,
     pub last_withdraw_request_shares: u128, // get zero as 0 when not in escrow
-    pub last_withdraw_request_value: u128,
+    pub last_withdraw_request_value: u64,
     pub last_withdraw_request_ts: i64,
 }
